Add CLI option to ignore files when linting
This unifies the exclusion of files across pep8 and pylint. Especially pylint is lacking in this departament. The current approach breaks if there as a huge number of files to process (command line too long). I got this error with 3000 files. Solution is to do it in batches but that change needs more planning as there are other bonuses possible with it (and drawbacks)