If linter config doesnt exist do not pass it
(change) From now on the linter tool will not try to use the config if
it doesn not exist. Before running the tool it will check if the given
(or default) config file exists and if not it won't pass the config flag
to the appropriate tool command. This is to avoid crashes when someone
just wants to run the tool with the default configuration.