Import commands so they appear as tested
Those modules artificially lower the coverage report stats. The code here will probably never be tested as those are pure usages of the click library. Testing this code would be testing click and not peltak. Just importing them so they don't appear as 0% as if it's not tested (should be 100% but let's settle at whatever only importing them gives us).