Use python3.6 docker image for CI
Since we're not testing on anything less than python3.6 we do not use
tox anymore. Hence we don't need a docker image with all the versions
of python (including 2). Python 3 is backwards compatible, so we can
just test on python3.6 and every python 3.6+ is supported out of the
box. We will support all LTS versions curretly supported by PSF.