Initial support for multiple version files
(feature) Support for multiple version files. The config now accepts
a list of files. The first file on the list will be treated as the source of
truth, but all files specified on the list will be written on any change.
This allows you to keep the version in the source files and configuration
synced.

(change) The version.file config variable is deprecated and will be
removed in the future versions. Use version.files and pass just
one file in the array. In the transition period, the version.file will be
converted to version.files by peltak internally, but this code path will
be removed with time.