Convert test core.util into a module
We have too many test modules and packages. It was supposed
to help readability but it actually harms it. We can use classes
to separate tests based on what they test instead of splitting
everything to it’s own module. If we get large modules this way
we can then split them up, but for the most tests it should be
manageable.