Add flake8
Seems like it can replace pylint which is the slowest of the bunch.
Right now the check is probably the slowest step in the CI build.
Looks like flake8 can catch the same errors we use pylint for and
is much faster (like 10x).

Initially we’ll leave pylint enabled and we’ll see if it catches any
errors missed by flake8. If there’s not examples of that in a longer
period of time, we can safely remove pylint.