Replace config src_dir with python_paths
(change) src_dir has been deprecated in favor of the new
python paths config value. It makes it possible to specify multiple
paths instead of one and the name seems more natural from the
user point of view. A python developer should already know what
PYTHON_PATH is and understand the meaining of python_paths
right away.

Same as with version.file vs version.files, we will leave src_dir for now.
It will be internally appended to the python_paths during the transition
period. This code path will be removed in the future.