Fix peltak test script
Running doctests in one command with regular tests causes pytest
to generate __pycache__ files in the test data which screws up the e2e
tests (extra files that were not expected). The fix is to not run the
doctests as part of all. Seems logical as they are a completely different
kind of tests. They test the documentation, not the code and live in a
different place than the rest of the tests.