change: Move basic commands out of extra
_more: We’ll be getting rid of peltak.extra package and instead
move everything out into a plugin or merge it into the core.

Versioning and basic git commands are so essential they will
be shipped directly with peltak. No point making separate
plugins for those. Especially since the versioning logic needs
to be provided as part of peltak so plugins can use it easily.

We want to avoid the situation where one plugin depends on
another.