change: Rename peltak.commands to peltak.cli
_more: Shorter and fits more into the new architecture. This will break
a lot of things, so it’s probably one of the last moments to od this,
since the whole architecture is changing right now. Move to scripts
as being the core functionality already changes a lot and will probably
be the final shape for v1.0. The old way of defining scripts will be
dropped soon as well.