feature: Combine plugin and scripts CLIs
_more: Until now, the CLI generated from the scripts would overwrite the
ones defined in plugins. This was especially annoying for groups as a single
script could remove access to a whole CLI group provided by plugins.
From now on, the plugins CLIs will take priority and the scripts commands will
be merged with it. If the script and plugin commands CLI paths clash, we will
raise an error prompting the user to change the script name.